requirements.txt
mysql-connector-python==8.0.18
PyMySQL==0.9.3
python-telegram-bot==12.8